refactor: change pathes to pages and rename files by index
